Continue work on study room offline

+++ Cloudents.Command/CommandHandler/ApplyCouponCommandHandler.cs
@@ -39,9 +39,9 @@ namespace Cloudents.Command.CommandHandler
             user.ApplyCoupon(coupon, tutor);
 
 
-            var tutorPrice = tutor.Price.Price;
+            var tutorPrice = (double)tutor.Price.Price;
 
-            message.NewPrice = Coupon.CalculatePrice(coupon.CouponType, tutorPrice, coupon.Value);
+            message.NewPrice = Coupon.CalculatePrice(coupon.CouponType, tutorPrice, (double)coupon.Value);
 
         }
     }
