+++ Cloudents.Core/DTOs/StudyRoomDto.cs
@@ -33,7 +33,7 @@ namespace Cloudents.Core.DTOs
         public decimal? CouponValue { get; set; }
 
         [EntityBind(nameof(StudyRoom.Price), nameof(Tutor.Price))]
-        public decimal TutorPrice { get; set; }
+        public double TutorPrice { get; set; }
         public string Jwt { get; set; }
 
         [EntityBind(nameof(BroadCastStudyRoom.BroadcastTime))]
